fixed: XLS export fails with more than 32767 rows

+++ sqlexplorer/src/net/sourceforge/sqlexplorer/sqleditor/results/export/ExporterXLS.java
@@ -55,7 +55,7 @@ public class ExporterXLS implements Exporter {
 		
 		Sheet sheet = wb.createSheet("sheet1");
 		
-		short rowNum = 0;
+		int rowNum = 0;
 
         if (includeColumnNames) {
     		Row row = sheet.createRow(rowNum++);
